Changed ID to long

+++ fact-manager-core/src/main/java/com/biit/factmanager/core/providers/FormRunnerFactProvider.java
@@ -19,7 +19,7 @@ public class FormRunnerFactProvider {
         this.formRunnerFactRepository = formRunnerFactRepository;
     }
 
-    public FormRunnerFact get(Integer factId) {
+    public FormRunnerFact get(Long factId) {
         return formRunnerFactRepository.findById(factId).orElseThrow(
                 () -> new FactNotFoundException(this.getClass(), "No fact with id '" + factId + "' found."));
     }
